Update MainActivity.java
This was a preliminary attempt to utilize Android Studio. After finally installing the software and virtual Android Device, there was a compiling error where a class had a slash symbol in the name, but we couldn't add a slash to the file name. For this reason, we decided to put a hold on this software and take our project to the web instead!